[Feat] Seperate Flower & Pill definition
Seperate class Flower and class Pill definition to Flower.swift and Pill.swift file
